Creating weapon mastery medal names dictionary (#105)
* Creating weapon mastery medal names dictionary

This adds a new folder to the /dictionaries directory for weaponMastery.  A single new file has been added for "medalNames.json" to provide names for the medal ids provided in weapon mastery.

* Updating MedalFirstBlood identifier

The API returns MedalFirstBlood but the image asset is MedalFirstblood.  Updating to respect the API response.

* Updating MedalAnnihilation

Spelling was off on the original.